Fix Postgres type mismatch crashing per-area blank_ballots writes
results_election_area.blank_ballots is a Postgres integer (INT4) column,
but insert_results_election_area_documents bound it as Option<i64>
(INT8). tokio-postgres rejects i64 values against an INT4-typed
parameter at bind time, so any tally with a non-empty blank ballot count
failed with "error serializing parameter 7" and retried until exhaustion
instead of writing results.

Only this one write path was affected: results_election.rs's inserts and
results_election_area.rs's insert_many_results_elections_areas both go
through jsonb_to_recordset, letting Postgres cast the JSON number itself
rather than relying on a Rust-typed bind.

Parent issue: https://github.com/sequentech/meta/issues/12773

+++ packages/windmill/src/postgres/results_election_area.rs
@@ -61,7 +61,7 @@ pub async fn insert_results_election_area_documents(
     area_id: &str,
     area_name: &str,
     documents: &ResultDocuments,
-    blank_ballots: Option<i64>,
+    blank_ballots: Option<i32>,
     blank_ballots_percent: Option<f64>,
 ) -> Result<()> {
     let documents_value = serde_json::to_value(documents.clone())?;
